:recycle: feat(event): EventService 추상화
