+++ src/main/java/com/kuku9/goods/domain/event/entity/Event.java
@@ -1,6 +1,7 @@
 package com.kuku9.goods.domain.event.entity;
 
 import com.kuku9.goods.domain.file.entity.File;
+import com.kuku9.goods.global.common.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,12 +12,17 @@ import jakarta.persistence.OneToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.SQLDelete;
+import org.hibernate.annotations.SQLRestriction;
+import org.hibernate.annotations.Where;
 
 @Getter
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
-public class Event {
+@SQLDelete(sql = "UPDATE event SET deleted_at=CURRENT_TIMESTAMP where event_id=?")
+@SQLRestriction("deleted_at IS NULL")
+public class Event extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
